+++ R/document.R
@@ -1,4 +1,3 @@
-
 #' Document (Chunks of) an R Code File
 #'
 #' @inheritParams fake_package
@@ -28,19 +27,21 @@
 #'     \item{pdf_path}{The path to the pdf file produced,}
 #'     \item{txt_path}{The path to the text file produced,}
 #'     \item{html_path}{The path to the html file produced,}
-#'     \item{check_result}{A list giving the \command{R CMD check} results.}
+#'     \item{check_result}{The return value of
+#'     \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}
 #' }
 #' @export
 #' @examples
 #' \donttest{
-#' res <- document(file_name = system.file("tests", "files", "minimal.R",
+#' res <- document(file_name = system.file("files", "minimal.R",
 #'                                         package = "document"),
 #'                 check_package = FALSE) # this is for the sake of CRAN cpu
 #'                 # time only. _Always_ stick with the default!
 #'
-#' # View R CMD check results.
-#' cat(res[["check_result"]][["stdout"]], sep = "\n")
-#' cat(res[["check_result"]][["stderr"]], sep = "\n")
+#' # View R CMD check results. If we had set check_package to TRUE in the above
+#' # example, we now could retrieve the check results via:
+#' cat(res[["check_result"]][["output"]][["stdout"]], sep = "\n")
+#' cat(res[["check_result"]][["output"]][["stderr"]], sep = "\n")
 #'
 #' # Copy docmentation to current working directory.
 #' # This writes to your disk, so it's disabled.
@@ -53,7 +54,7 @@ document <- function(file_name,
                      output_directory = tempdir(),
                      dependencies = NULL, sanitize_Rd = TRUE, runit = FALSE,
                      check_package = TRUE, check_as_cran = check_package,
-                     stop_on_check_not_passing = TRUE, clean = FALSE,
+                     stop_on_check_not_passing = check_package, clean = FALSE,
                      debug = TRUE, ...) {
     if (is.null(working_directory))
         working_directory <- file.path(tempdir(),
@@ -88,15 +89,15 @@ document <- function(file_name,
     return(status)
 }
 
-#' Read R Documentation Files from a Package's Source, Convert and Write Them
-#' to Disk
+#' Write Documentation to Disk
 #'
-#' What does it?
+#' Read R documentation files from a package's source, convert and write them
+#' to disk.
 #'
 #' \code{file_name} will usually be provided by \code{\link{document}} as the
 #' R code file's name. This may, differing from a (temporary) package's name,
 #' contain underscores. If you use the functions directly: stick with the
-#' default, in which case the internally
+#' default, in which case internally
 #' the  \code{\link[base]{basename}} of your \code{package_directory} will be
 #' used. This should be a good guess.
 #' @inheritParams fake_package
@@ -115,6 +116,7 @@ document <- function(file_name,
 #'     \item{txt_path}{The path to the text file produced.}
 #'     \item{html_path}{The path to the html file produced.}
 #' }
+#' @keywords internal
 write_the_docs <- function(package_directory, file_name = package_directory,
                            output_directory = tempdir(),
                            dependencies = NULL, sanitize_Rd = TRUE,
@@ -143,11 +145,12 @@ write_the_docs <- function(package_directory, file_name = package_directory,
     }
     if (! dir.exists(output_directory)) dir.create(output_directory)
     options("document_package_directory" = package_directory)
-    call_pdf <- callr::rcmd_safe("Rd2pdf",
-                                 c("--no-preview --internals --force",
-                                   paste0("--title=", pdf_title),
-                                   paste0("--output=", pdf_path),
-                                   man_directory))
+    rcmd_args <- c("--no-preview", "--internals", "--force",
+                   paste0("--title=", pdf_title), paste0("--output=", pdf_path),
+                   man_directory)
+    call_pdf <- withr::with_dir(tempdir(),
+                                callr::rcmd_safe("Rd2pdf", rcmd_args)
+                                )
     if (! as.logical(call_pdf[["status"]])) status[["pdf_path"]]  <- pdf_path
     files  <- sort_unlocale(list.files(man_directory, full.names = TRUE))
     # using R CMD Rdconv on the system instead of tools::Rd2... since