+++ inst/doc/Introduction_to_document.R
@@ -1,5 +1,5 @@
 ## ---- comment = ""-------------------------------------------------------
-path <- system.file("tests", "files", "minimal.R", package = "document")
+path <- system.file("files", "minimal.R", package = "document")
 cat(readLines(path), sep = "\n")
 
 ## ---- result = "hide", message = FALSE-----------------------------------
@@ -12,12 +12,12 @@ cat(readLines(d[["txt_path"]]), sep = "\n")
 project_root <- rprojroot::find_root(rprojroot::is_r_package)
 d <- document::document(file_name = path,
                         output_directory = file.path(project_root,
-                                                     "inst", "tests", "files"),
+                                                     "inst", "files"),
                         check_package = FALSE)
 file.remove(unlist(d[c("txt_path", "pdf_path")]))
 
 ## ---- comment = ""-------------------------------------------------------
-path <- system.file("tests", "files", "simple.R", package = "document")
+path <- system.file("files", "simple.R", package = "document")
 cat(readLines(path), sep = "\n")
 
 ## ---- result = "hide", message = FALSE-----------------------------------
@@ -37,6 +37,6 @@ pager_function <- function(x, ...) return(cat(paste(readLines(x),
 options(pager = pager_function)
 
 ## ---- comment = "", message = FALSE, warning = FALSE---------------------
-path <- system.file("tests", "files", "minimal.R", package = "document")
+path <- system.file("files", "minimal.R", package = "document")
 document::man(x = path, topic = "foo")