+++ inst/tests/testthat.R
@@ -1,3 +1,0 @@
-library("testthat")
-library("document")
-test_check("document")